refactor: ProfileRootView와 ProfileView로 분리
