+++ Projects/Feature/Profile/Interface/Sources/Profile/Secondary/ProfileHeartRateView.swift
@@ -10,9 +10,9 @@ import ComposableArchitecture
 import SharedDesignSystem
 
 struct ProfileHeartRateView: View {
-    private let store: StoreOf<ProfileRootStore>
+    private let store: StoreOf<ProfileStore>
 
-    init(store: StoreOf<ProfileRootStore>) {
+    init(store: StoreOf<ProfileStore>) {
         self.store = store
     }
